requirements.txt
beautifulsoup4==4.4.0
blinker==1.3
cffi==1.1.2
click==4.1
coverage==3.7.1
cryptography==0.9.3
decorator==3.4.2
enum34==1.0.4
Flask==0.11.dev0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
idna==2.0
ipaddress==1.0.14
itsdangerous==0.24
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
mdx-twitter==0.2.6
ndg-httpsclient==0.4.0
oauthlib==0.7.2
pbr==1.3.0
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
python-openid==2.2.5
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
SQLAlchemy==1.0.6
sqlalchemy-migrate==0.9.7
sqlparse==0.1.15
Tempita==0.5.2
Werkzeug==0.10.4
Whoosh==2.7.0
wikipedia==1.4.0
WTForms==2.0.2
